feat: 마이페이지 밸런스게임 데이터 변환 로직을 유틸리티 함수로 추가합니다.
